renamed package ui.main to ui.onboarding
