+++ app/src/main/res/layout-v26/fragment_onboarding_support.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.PlaceholderFragment">
+    tools:context=".ui.onboarding.PlaceholderFragment">
 
     <ImageView
         android:id="@+id/onboarding_imageView"
